+++ OSTC/src/trajectory.cpp
@@ -18,11 +18,11 @@ bool SamplePoint::operator==(const SamplePoint& other) const
 
 Trajectory::Trajectory(const uint32_t id, const std::vector<SamplePoint>& points): id(id), points(points) {}
 
-Trajectory::Trajectory(const uint32_t id, const std::vector<SamplePoint>& points, const short start_index,
-                       const short end_index): id(id), points(points), start_index(start_index), end_index(end_index)
+Trajectory::Trajectory(const uint32_t id, const std::vector<SamplePoint>& points, const int start_index,
+                       const int end_index): id(id), points(points), start_index(start_index), end_index(end_index)
 {}
 
-Trajectory Trajectory::operator()(const short start, const short end)
+Trajectory Trajectory::operator()(const int start, const int end)
 {
     if (end + 1 > points.size()) {
         return Trajectory(id, std::vector<SamplePoint>(points.begin() + start, points.begin() + end), start, end);
@@ -51,7 +51,7 @@ bool ReferenceTrajectory::operator==(const ReferenceTrajectory& other) const
 
 std::size_t std::hash<Trajectory>::operator()(const Trajectory& t) const noexcept
 {
-    return ((hash<uint32_t>()(t.id) ^ (hash<short>()(t.start_index) << 2) ^ (hash<short>()(t.end_index) >> 1) ^
+    return ((hash<uint32_t>()(t.id) ^ (hash<int>()(t.start_index) << 2) ^ (hash<int>()(t.end_index) >> 1) ^
              (hash<uint32_t>()(t.points.size()) << 1)) >>
             1);
 }
@@ -150,7 +150,7 @@ OSTCResult Trajectory::OSTC(std::unordered_map<Trajectory, std::vector<Trajector
     }
 
     std::vector<int> Ft(points.size() + 1, 0);      // +1 for F_T[0] = 0
-    std::vector<short> pre(points.size() + 1, -1);  // -1 indicates no predecessor
+    std::vector<int> pre(points.size() + 1, -1);  // -1 indicates no predecessor
     std::vector<ReferenceTrajectory> T_prime;
 
     for (size_t i = 1; i <= points.size(); ++i) {
